requirements-build.txt
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
pandas==2.2.3
pyarrow==18.1.0
db-dtypes==1.3.1
